@@ -2,8 +2,6 @@
   "name": "episodic",
   "version": "0.0.0-development",
   "description": "An opinionated CLI tool to automatically rename TV show releases",
-  "author": "Rafael Bardini",
-  "homepage": "https://github.com/rbardini/episodic",
   "keywords": [
     "tv show",
     "series",
@@ -13,41 +11,45 @@
     "imdb",
     "cli"
   ],
-  "main": "index.js",
-  "scripts": {
-    "format": "prettier --write *.js bin/*",
-    "lint": "eslint *.js bin/*",
-    "test": "jest"
+  "homepage": "https://github.com/rbardini/episodic",
+  "bugs": {
+    "url": "https://github.com/rbardini/episodic/issues"
   },
-  "bin": "./bin/episodic",
   "repository": {
     "type": "git",
     "url": "https://github.com/rbardini/episodic.git"
   },
-  "bugs": {
-    "url": "https://github.com/rbardini/episodic/issues"
-  },
   "license": "MIT",
+  "author": "Rafael Bardini",
+  "type": "module",
+  "exports": "./index.js",
+  "bin": "./bin/episodic.js",
+  "files": [
+    "./bin/*",
+    "./index.js"
+  ],
+  "scripts": {
+    "format": "prettier --ignore-path .gitignore .",
+    "lint": "eslint --ignore-path .gitignore .",
+    "test": "FORCE_COLOR=0 vitest run --coverage"
+  },
   "dependencies": {
-    "chalk": "^4.1.0",
+    "chalk": "^5.0.0",
     "commander": "^9.0.0",
     "imdb-api": "^4.4.0",
     "inquirer": "^8.1.0",
     "lodash.groupby": "^4.6.0",
-    "ora": "^5.4.0",
+    "ora": "^6.0.0",
     "video-name-parser": "^1.4.0"
   },
   "devDependencies": {
-    "eslint": "7.29.0",
-    "eslint-config-standard": "16.0.3",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "5.1.1",
-    "jest": "27.5.1",
+    "c8": "7.11.3",
+    "eslint": "8.17.0",
     "prettier": "2.6.2",
-    "strip-ansi": "6.0.0"
+    "prettier-plugin-packagejson": "2.2.18",
+    "vitest": "0.14.2"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14.16"
   }
 }
